V0.9.0a “Explorer” (#164)
# V0.9.0a “Explorer” release notes
## Release date: June 4th, 2018

### The Explorer release
Explorer now supports more features on the privacy layer which includes the zero-knowledge algorithm published a month ago in Zond. It brings the ability to handle private data within the system in such a way that the owner can retain control of the information by their DC (data creator) node, while publishing cryptographic commitments in the system to the DH (data holder) nodes involved in replication. 

Our data importer now supports the "Web of Things" (WoT) standard developed by the  World Wide Web Consortium (W3C). It is a first major step towards IoT support and one of the major requirements as a consequence of the first use cases, our partners and community. 

## New features:
- Improved privacy layer implementation 
- Web of Things model supported by our data importer
- New version of Houston (GUI) with data graph visualisation and improved interface
- Remote control API improvement
- IoC implementation in the code

## Fixes: 
- Importer error handling fixes
- Better organisation of the code and refactoring of the singletone classes - IoC
- Improved blockchain tests - truffle test suite
- Additional refactoring of the code
- Increased test coverage